modules/export-to-bq-incremental/function-source-code/requirements.txt
functions-framework==3.*
requests==2.32.3
google-auth==2.38.0
google-cloud-bigquery==3.29.0